Rename receive-message as campaignActivity
