+++ app/routes/index.js
@@ -3,11 +3,11 @@
 // Routes
 const campaignsIndexRoute = require('./campaigns/index');
 const campaignsSingleRoute = require('./campaigns/single');
-const receiveMessageRoute = require('./receive-message');
+const campaignActivityRoute = require('./campaignActivity');
 const statusRoute = require('./status');
 
 // middleware config
-const authenticateConfig = require('../../config/middleware/authenticate');
+const authenticateConfig = require('../../config/lib/middleware/authenticate');
 
 // middleware
 const timeoutMiddleware = require('../../lib/middleware/timeouts');
@@ -35,5 +35,8 @@ module.exports = function init(app) {
   app.use('/v1/campaigns', campaignsIndexRoute);
 
   // Receives inbound message from Gambit Conversations service.
-  app.use('/v1/receive-message', receiveMessageRoute);
+  app.use('/v1/campaignActivity', campaignActivityRoute);
+  // TODO: Remove this once we make the switch in Conversations.
+  // TODO: Remove this once we make the switch in Conversations.
+  app.use('/v1/receive-message', campaignActivityRoute);
 };
