+++ config/lib/middleware/authenticate.js
