--- config/middleware/campaigns/required-params.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const configVars = {};
-
-configVars.containerProperty = 'body';
-configVars.requiredParams = [
-  'phone',
-  'type',
-];
-module.exports = configVars;
