--- config/middleware/signups/required-params.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const configVars = {};
-
-configVars.containerProperty = 'body';
-configVars.requiredParams = [
-  'id',
-  'source',
-];
-module.exports = configVars;
