+++ documentation/README.md
@@ -4,4 +4,4 @@ Endpoint                                       | Functionality
 ---------------------------------------------- | --------------------------------------------------------
 `GET /v1/campaigns` | [Retrieve all Campaigns with keywords available in Gambit](endpoints/campaigns.md#retrieve-all-campaigns)
 `GET /v1/campaigns/:id` | [Retrieve a single Campaign](endpoints/campaigns.md#retrieve-a-campaigns)
-`POST /v1/receive-message` | [Receive a Campaign Signup Message](endpoints/receive-message.md)
+`POST /v1/campaignActivity` | [Receive a Campaign Signup Message](endpoints/campaignActivity.md)
