+++ lib/middleware/campaignActivity/draft-caption.js
