+++ lib/middleware/campaignActivity/draft-photo.js
