+++ lib/middleware/campaignActivity/menu-completed.js
