+++ lib/middleware/campaignActivity/signup-get.js
