+++ lib/middleware/campaignActivity/text-post.js
